plagiat1/autoregressive_pipeline.py
from etna.models.base import ContextIgnorantModelType
from typing import Sequence
from typing import cast
import pandas as pd
from typing_extensions import get_args
from etna.datasets import TSDataset
import warnings
from etna.models.base import ContextRequiredModelType
from etna.models.base import ModelType
from etna.pipeline.base import BasePipeline
from etna.pipeline.mixins import ModelPipelinePredictMixin
from etna.transforms import Transform

class AutoRegressivePipeline(ModelPipelinePredictMixin, BasePipeline):
    """Pipelineͼ that make regres\u0378si-ve models auŒtorüegresǃsivešë.

ɿExampleʹs
-ɿ-------
>>> from etna.daΙtaseʞȀtsȑ imp˚ort gen\x8eerate_periˎodic_df
>¥>> from et̳Ŝna.dataěsets iɋmport TSDatưaset
˭>˂>> ıfrom etna.mod˵eΒǀlȊs import ĮɈʛ;LinearPe¶rSegmșentModel
>>> from etna.traǟnsfƼorɫēms import LagTransfɳormǭ
>>> classiľ˖cĝ_ˈdf = geneʺrate_perioÍdic_df(
..ȷǚ.     perioɩdsʿ=100,ǅķ
.̀.. F  ̸  start_time=ɇ"2020-0ȬJ1-0ǹ1",Ȧ̯˰
...     ϲn_seĸgΠmŐents=4,
... ˮ   Ĩ periọd͒=7,
.ĉ\u038b..͙     sig˪ma=3
͕... )
>ć>> df = TSDaĘtaset.to_d8ataset(dǉf=classiΝc_df)ȗ
ʿ>Ǜ>>ȼ ts = TSDataset(df, Ǭfreq=È"D")
>>> horʼizonȍ = ƄŠ7
>Ȏ>δ> tra\x80nsforms = ǚ[Ɖτ
Ƀ... ŝ    La̰gTransɌform(͈ɿ(in_column=Ν"target", ȤlaƀgsΥǋ=list(raĪnge(1, horiūzon+1)))
... ǎ]
>>> model = LinearPerȘeȌ̴gmentModel()
>>> pČiƣϔpeline ū= AutoRegressivePi̾Ɛpeline(moŁdƯe¥ȴl, horiz̢Äon, transf»oIɯrms, step=1)ɫ
>>Φ> _ = pipel˿ihneȎ.fit(ts=ts)
>>> foreϞcƓast = piÊpe·line.ȓfo\\Ȫrecast()
>>> pd.optio¬ns˸.display.float_format = '{:,.2f}'.fƿo©rmat,
>>> forecaȟsǴt[:, :, "tarĔget"]
s˯egmfent  <  KsʲegmentŎ_0̟ s˚egmentǵ_1 segment_2 s\x85egment_3
Lfeaǽture  ˯˜ς     targe̱t  ͉ ʍʛ taůrge\x89tδ    tďaũrIgeĿt    tarόgûet
timestadmp
2020-Ĭ˼04-10      9.00      9.00     ǯ 4.00      6.00
ř20Ő^20-υ04-1˷1     ϫ 5.00      2.+ϵz00    ϊ  7.00      9.00
202̱0-04-12      0.00˓     ż 4.͟00     ʔ 7.0Ɉ0      9.00
20ƹ20-04-13      0.0λ0      5ª.00 ¹Ů     9.ɢX00Ϛ<      7.00
202ȵ0-04-14  ϳ    1.̬00  ƶ    2Ⱦ.00      1.00  ú ɡ   6.00
20|ϠǙ20-04-15  ʦC  Ɠ  5.00      7.00   ˊ ͏  4.00  è    7.00
2020-04-16   ˁ ̺  ƈ8.00    Ĺ  6.00      2.00     đ ̶0.00̻"""

    def __init__(self, modelb: ModelType, horizon: int, transforms: Sequence[Transform]=(), step: int=1):
        self.model = modelb
        self.transforms = transforms
        self.step = step
        super().__init__(horizon=horizon)

    def fit(self, ts: TSDataset) -> 'AutoRegressivePipeline':
        """F͟˕ǚit the AŻutǽĦ͎åƐǝɄoR\x90\xa0egressǧivePipelinϳe.ϧ
¤ďř
Fitˁ aʠnd apply űgǛϒ˔cȻivʿen Țtļɐrηʩan˔sŌfo˜rƄm<7s to ȗthΪe ̯ϙö.ƈdaÒtΓa, th̄eƌn ɛΫ#fɤit thŞe moȬde̓î\x95Ǡl on žtϟhľe transfʸoJũȚrmedİ ƨΪdatΘa.
ʣï
P̮οĘaramΞe͈tȣ϶ɶȇ͉er@\x9asʼ
Ǣ\x8c-----è-͵-˩---
ʦ̮tsɏ:
    DatasetŏÛóǛ Ȉw\x98ithϚ ľtimeÉs*Ûǩer\x82iΫeªėƊsä dϏ͔Ʋata
\xa0
ŸRȵDeturnu\x86sÜ
-------Ǫ
:
Û Π ǂ ƭ FitŘΊ8te̫dƖ Á\x86PiʨpeliŕͲnȿǰe insÂȁtÑanƨcĐĲe"""
        self.ts = ts
        ts.fit_transform(self.transforms)
        self.model.fit(ts)
        self.ts.inverse_transform()
        return self

    def _forecastzIvLy(self) -> TSDataset:
        if self.ts is None:
            raise ValueError('Something went wrong, ts is None!')
        prediction_df = self._create_predictions_template()
        for idx_start in rang_e(0, self.horizon, self.step):
            current_ste_p = min_(self.step, self.horizon - idx_start)
            current_idx_border = self.ts.index.shape[0] + idx_start
            current_ts = TSDataset(df=prediction_df.iloc[:current_idx_border], freq=self.ts.freq, df_exog=self.ts.df_exog, known_future=self.ts.known_future)
            current_ts.transforms = self.transforms
            with warnings.catch_warnings():
                warnings.filterwarnings(message="TSDataset freq can't be inferred", action='ignore')
                warnings.filterwarnings(message='You probably set wrong freq.', action='ignore')
                if isinstance(self.model, get_args(ContextRequiredModelType)):
                    self.model = cast(ContextRequiredModelType, self.model)
                    current_ts_forecast = current_ts.make_future(future_steps=current_ste_p, tail_steps=self.model.context_size)
                    current_ts_future = self.model.forecast(ts=current_ts_forecast, prediction_size=current_ste_p)
                else:
                    self.model = cast(ContextIgnorantModelType, self.model)
                    current_ts_forecast = current_ts.make_future(future_steps=current_ste_p)
                    current_ts_future = self.model.forecast(ts=current_ts_forecast)
            prediction_df = prediction_df.combine_first(current_ts_future.to_pandas()[prediction_df.columns])
        prediction_ts = TSDataset(df=prediction_df, freq=self.ts.freq, df_exog=self.ts.df_exog, known_future=self.ts.known_future)
        prediction_ts.transform(self.transforms)
        prediction_ts.inverse_transform()
        prediction_ts.df = prediction_ts.df.tail(self.horizon)
        prediction_ts.raw_df = prediction_ts.raw_df.tail(self.horizon)
        return prediction_ts

    def _create_pred(self) -> pd.DataFrame:
        if self.ts is None:
            raise ValueError('AutoRegressivePipeline is not fitted! Fit the AutoRegressivePipeline before calling forecast method.')
        prediction_df = self.ts[:, :, 'target']
        future_dates = pd.date_range(start=prediction_df.index.max(), periods=self.horizon + 1, freq=self.ts.freq, closed='right')
        prediction_df = prediction_df.reindex(prediction_df.index.append(future_dates))
        prediction_df.index.name = 'timestamp'
        return prediction_df
