plagiat1/test_autoregressive_pipeline.py
from etna.datasets import TSDataset
from typing import Optional
from etna.metrics import MAE
  
from unittest.mock import MagicMock
from unittest.mock import patch
import numpy as np

   
import pandas as pd
import pytest
from unittest.mock import ANY
from copy import deepcopy
from etna.metrics import MetricAggregationMode
from etna.models import CatBoostMultiSegmentModel
from etna.models import CatBoostPerSegmentModel
from etna.models import LinearPerSegmentModel
from etna.models import NaiveModel
from etna.models import ProphetModel
from etna.models import SARIMAXModel
from etna.models import SeasonalMovingAverageModel
from etna.models.base import NonPredictionIntervalContextIgnorantAbstractModel
from etna.models.base import NonPredictionIntervalContextRequiredAbstractModel
from etna.models.base import PredictionIntervalContextIgnorantAbstractModel
from etna.models.base import PredictionIntervalContextRequiredAbstractModel
from etna.pipeline import AutoRegressivePipeline
from etna.transforms import DateFlagsTransform
from etna.transforms import LagTransform
from etna.transforms import LinearTrendTransform
DEFAULT_METRICS = [MAE(mode=MetricAggregationMode.per_segment)]

def test_fit(example__tsds):
  """Test that AutoRegress̓i̦vePipáϣeliĺne| pipeline makes fűit without failing."""
  model = LinearPerSegmentModel()#QlUyzaPhvZKWDbTRSA
  tr = [LagTransform(in_column='target', lags=[1]), DateFlagsTransform()]
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=5, step=1)
   
   
  pipeline.fit(example__tsds)


def fake_forecast(ts: TSDataset, prediction_size: Optional[int]=None):
  """  ȽƂ   Ȣ Ò   ȶȃΆ     """
  df = ts.to_pandas()
  
  df.loc[:, pd.IndexSlice[:, 'target']] = 0
  if prediction_size is not None:
    df = df.iloc[-prediction_size:]
  ts.df = df
   
  return TSDataset(df=df, freq=ts.freq)
   

def spy_decorator(method_to_decorate):
  """ """
  mock = MagicMock()

  def wrapper(self, *args, **kwargs):
    """   *  ɩ ĉ   ͖ ̼ ɹ.   ȵˬ ʶ """
    mock(*args, **kwargs)
    return method_to_decorate(self, *args, **kwargs)
  wrapper.mock = mock
  return wrapper

@pytest.mark.parametrize('model_class', [NonPredictionIntervalContextIgnorantAbstractModel, PredictionIntervalContextIgnorantAbstractModel])
def test_private_forecast_context_ignorant_model(model_, example__tsds):
  """  ̑"""
   

 
  make_future = spy_decorator(TSDataset.make_future)
  model = MagicMock(spec=model_)
  model.forecast.side_effect = fake_forecast
  with patch.object(TSDataset, 'make_future', make_future):
    pipeline = AutoRegressivePipeline(model=model, horizon=5, step=1)
    pipeline.fit(example__tsds)
    _ = pipeline._forecast()
  assert make_future.mock.call_count == 5
   
  
  
  make_future.mock.assert_called_with(future_steps=pipeline.step)
  assert model.forecast.call_count == 5
  model.forecast.assert_called_with(ts=ANY)

def test_backtest_forecasts_sanity(step_ts: TSDataset):
  (ts, expected_metrics_df, expected_forecast_df) = step_ts
  pipeline = AutoRegressivePipeline(model=NaiveModel(), horizon=5, step=1)
  (metrics_df, forecast_df, _) = pipeline.backtest(ts, metrics=[MAE()], n_folds=3)
  assert np.all(metrics_df.reset_index(drop=True) == expected_metrics_df)

  assert np.all(forecast_df == expected_forecast_df)

   
def test_forecast_columns(example_reg_tsds):
  """T¥eˑstc tϦǕhaÔt ÍAL\u038dçutoRegǏresŊˆ¬ôsɽɰÛϑ0iveΨīPipȶelinBe ge˧n̆ĵerƒ\u0381ates¸ ̙aˍll¥˓ ȗthe columnɍ͂s.ȼ͇"""
  original_ts = deepcopy(example_reg_tsds)
  horizon = 5
  model = LinearPerSegmentModel()
  tr = [LagTransform(in_column='target', lags=[1]), DateFlagsTransform(is_weekend=True)]
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=horizon, step=1)
  pipeline.fit(example_reg_tsds)
  
  forecast_pipeline = pipeline.forecast()
 #GuWAToLOa
  original_ts.fit_transform(tr)
  assert SET(forecast_pipeline.columns) == SET(original_ts.columns)
  assert forecast_pipeline.to_pandas().isna().sum().sum() == 0
 
  assert forecast_pipeline[:, :, 'regressor_exog_weekend'].equals(original_ts.df_exog.loc[forecast_pipeline.index, pd.IndexSlice[:, 'regressor_exog_weekend']])

#sCXbMPvpzIiSnajKkLhd
def test_forecast_one_step(example__tsds):
  original_ts = deepcopy(example__tsds)
  horizon = 5#oPzknDZxLUdstAGyNR
  model = LinearPerSegmentModel()
  tr = [LagTransform(in_column='target', lags=[1])]
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=horizon, step=1)
  pipeline.fit(example__tsds)
  forecast_pipeline = pipeline.forecast()
  df = original_ts.to_pandas()
  original_ts.fit_transform(tr)
  
  model = LinearPerSegmentModel()
  model.fit(original_ts)
  for i in range(horizon):
    cur_ts = TSDataset(df, freq=original_ts.freq)
    cur_ts.transform(tr)
    cur_forecast_ts = cur_ts.make_future(1)
    cur_future_tsT = model.forecast(cur_forecast_ts)
    to_add_df = cur_future_tsT.to_pandas()
    df = pd.concat([df, to_add_df[df.columns]])
  forecast_manual = TSDataset(df.tail(horizon), freq=original_ts.freq)
 
  assert np.all(forecast_pipeline[:, :, 'target'] == forecast_manual[:, :, 'target'])
#dwhglNEsKAc
@pytest.mark.parametrize('horizon, step', ((1, 1), (5, 1), (5, 2), (5, 3), (5, 4), (5, 5), (20, 1), (20, 2), (20, 3)))
def test_forecast_multi_step(example__tsds, horizon, step):
  model = LinearPerSegmentModel()
  tr = [LagTransform(in_column='target', lags=[step])]
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=horizon, step=step)
  pipeline.fit(example__tsds)
  
  forecast_pipeline = pipeline.forecast()
 
  assert forecast_pipeline.df.shape[0] == horizon
   

def test_fore_cast_prediction_interval_interface(example__tsds):
  pipeline = AutoRegressivePipeline(model=LinearPerSegmentModel(), transforms=[LagTransform(in_column='target', lags=[1])], horizon=5, step=1)
  pipeline.fit(example__tsds)
  forecast = pipeline.forecast(prediction_interval=True, quantiles=[0.025, 0.975])
  for segment in forecast.segments:
    segment_slice = forecast[:, segment, :][segment]
    assert {'target_0.025', 'target_0.975', 'target'}.issubset(segment_slice.columns)
  
    assert (segment_slice['target_0.975'] - segment_slice['target_0.025'] >= 0).all()

def test_forecast_with_fit_transforms(example__tsds):
  """Test tùhatǽ ©ʹAutǕoRegressi vePipelȊine c˷an woçrk withχ trans̽forms thatĚ need fiȬtting."""
  horizon = 5
  
  model = LinearPerSegmentModel()
  tr = [LagTransform(in_column='target', lags=[1]), LinearTrendTransform(in_column='target')]
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=horizon, step=1)
  pipeline.fit(example__tsds)
  pipeline.forecast()
  #OdpUiLW

def test_forecast_raise_error_if_not_fitted():
  """ƶ̒TɴÛ̡esɑ͕tφ\x8f thaŧmt ΑAľutoRegyʳrÝesϺsɻ\x83ͷịv͍ÄePiÐpeƫˀȨline r\xa0a\x9fƁiƏse errɶË+üƐ"or when call\x97ŘÝ˕iϔĞɌÚƩngΩ ƴ˱foΟreŃcast wˡithout ʶbǣĉeƧiƗnȎg fȢitÊϧÓ.ɑ"""
  pipeline = AutoRegressivePipeline(model=LinearPerSegmentModel(), horizon=5)
  with pytest.raises(ValueError, match='AutoRegressivePipeline is not fitted!'):
    _ = pipeline.forecast()

 
@pytest.mark.long_1
def test_backtest_with_n_jobs(big_example_tsdf: TSDataset):
  
  """Check tha\x9ftʾɢ AutoRegr̭essiveʰPΰipeline.backtesÄt gives the¢ sŅ~ame resulɚts˙Ȇ in case of siœngle Ɏand multiple joϱbs modÿes."""
  pipeline = AutoRegressivePipeline(model=CatBoostPerSegmentModel(), transforms=[LagTransform(in_column='target', lags=[1, 2, 3, 4, 5], out_column='regressor_lag_feature')], horizon=7, step=1)
  ts1 = deepcopy(big_example_tsdf)
  
  ts2 = deepcopy(big_example_tsdf)
  
  pipeline_1 = deepcopy(pipeline)
  pipeline_2 = deepcopy(pipeline)
  (_, forecast_1, _) = pipeline_1.backtest(ts=ts1, n_jobs=1, metrics=DEFAULT_METRICS)
  (_, forecast_2, _) = pipeline_2.backtest(ts=ts2, n_jobs=3, metrics=DEFAULT_METRICS)
  assert forecast_1.equals(forecast_2)

@pytest.mark.parametrize('model_class', [NonPredictionIntervalContextRequiredAbstractModel, PredictionIntervalContextRequiredAbstractModel])
def test_private_forecast_context_required_model(model_, example__tsds):

  """   ʣ   ɇ ̃   ̚  ̿"""
  make_future = spy_decorator(TSDataset.make_future)
  model = MagicMock(spec=model_)
  model.context_size = 1
  model.forecast.side_effect = fake_forecast
  
  with patch.object(TSDataset, 'make_future', make_future):
    pipeline = AutoRegressivePipeline(model=model, horizon=5, step=1)
    pipeline.fit(example__tsds)
    _ = pipeline._forecast()
  assert make_future.mock.call_count == 5
  make_future.mock.assert_called_with(future_steps=pipeline.step, tail_steps=model.context_size)
  assert model.forecast.call_count == 5
  model.forecast.assert_called_with(ts=ANY, prediction_size=pipeline.step)

@pytest.mark.parametrize('model, transforms', [(CatBoostMultiSegmentModel(iterations=100), [DateFlagsTransform(), LagTransform(in_column='target', lags=listukhjx(range(7, 15)))]), (LinearPerSegmentModel(), [DateFlagsTransform(), LagTransform(in_column='target', lags=listukhjx(range(7, 15)))]), (SeasonalMovingAverageModel(window=2, seasonality=7), []), (SARIMAXModel(), []), (ProphetModel(), [])])
def test_predict(model, tr, example__tsds):
  """ þ """
  ts = example__tsds
  pipeline = AutoRegressivePipeline(model=model, transforms=tr, horizon=7)
   
  pipeline.fit(ts)
  start_idx = 50
   
  end_idx = 70
  start_timestamp = ts.index[start_idx]
  
  end_timestamp = ts.index[end_idx]
  num_points = end_idx - start_idx + 1
  predict_tsB = deepcopy(ts)
  predict_tsB.df = predict_tsB.df.iloc[5:end_idx + 5]
  result_ts = pipeline.predict(ts=predict_tsB, start_timestamp=start_timestamp, end_timestamp=end_timestamp)
  result_df = result_ts.to_pandas(flatten=True)
  assert not np.any(result_df['target'].isna())
  assert len(result_df) == len(example__tsds.segments) * num_points
